Fixed a bug and added print statements
Original code had a real issue in that I told it to use UTC and then one of the libraries I was using assumed I meant my system time, so I fixed it.  I also added a test event scheduled a minute from program launch and added print statements to inform the user that the program was doing things.